Convert the nuclear attack video into an audio file.
The attack.mp4 video is a recording of a 1961 nuclear attack warning
(which, thankfully, never had to be used). It was intended to be played
on the CONELRAD network. The video part is just a still image of the
CONELRAD logo. This video takes up over 130MB on the SD card when
encoded for playback on the badge, but only 11MB in audio form. I
think it makes more sense to have it as an audio file. Also it does a
good job of exercising the LED bargraph on the music player.