Add a radio remote unlocker app
A special unlock flag called UL_BLACKBADGE has been added. This
attribute can not be unlocked by the app-unlock puzzle app. It either
has to be set as a default bit in the userconfig at compile time,
or it has to be unlocked by phone or someone else with a black
badge.

Two new orchard app flags have been added: APP_FLAG_UNLOCK and
APP_FLAG_BLACKBADGE. This allows applications to be unlocked via
the puzzle unlocker. It also makes the firmware OTA sender app
and radio remote unlocker app visible in the launcher only to
black badge holders.

The radio remote unlocker has to send the access password before
it can change the unlock word. This is the same password in the
UICR that all the other badges use. This means the UICR must be
configured the same way on both the black badge and the badge
being unlocked in order for it to work.