Rework the ChibiOS nRF52 RNG driver
The existing RNG implementation spins waiting for a random byte to
be available. This causes problems when, for example, you set the
LED config to a mode that uses random numbers, disable the radio (which
will use the ChibiOS RNG instead of the SoftDevice) and then run the
bitcoin app which plays a video. It stutters like crazy.

The RNG module has an interrupt which signals when a byte is ready.
We can put the calling thread to sleep and wake up when the interrupt
fires instead of just busy-looping. This makes much more efficient
use of the CPU and avoids the stuttering problem during video playback.