Add a first attempt at a radio spectrum display app
This app uses a custom driver to directly access the NRF52 radio.
When it's running, the softdevice is disabled. It scans by sweeping
the radio across the 2400MHz to 2499MHz channels and taking an RSSI
reading at each point. It's not terribly flashy due to the limited
number of channels, but it's nominally functional.