Fix some issues with the "puzwatch" app
Spawning a separate thread just to monitor a GPIO pin is wasteful. Now
that we've adopted the diode logic trick for buttons, we're only using
one of the 8 available GPIOTE channels for button interrupts. That means
there's 7 available. We can use one more to monitor the SAO GPIO2
pin to trigger the puzwatch app. This means we'll get an interrupt on
the falling edge when the pin is pulled down, and the orchard app thread
can then launch the puzwatch app. We don't need to waste CPU cycles polling
the pin every few milliseconds. (It's bad enough we have to do that for
the touch panel.)

I've tested this with my nRF52840 DK board and it seems to work fine.

The puzwatch also has a memory leak: it forgets to destroy the ghPin
label widget on exit. This has been fixed.