Make a few game updates
Devon sent me his latest iteration of his sprite code. It implements
support for cacheing the game screen map background, which app-battle doesn't
use yet, but which it could later.

Fixed a bug in isp_check_sprites_collision(): the second call to
make_bounding_rect() was using the wrong background buffer pointer
(typo: instead of id2/id1, it should have been id2/id2)

Also, make the game reflect enemy movements on the world map. We
weren't changing the sprite state for old enemy discoveries. (This
seems a little slow... need to talk to John if this needs to be
fixed.)