Fix performance issue in music player
Moving the ledDraw() routine to led.c actually hurts performance
for the music player. We're pushing the envelope with regards to
how many spare cycles we have between playing audio samples. Drawing
the spectragraph and updating the LED array takes a fair amount of
math and work. With ledDraw() in the music app proper, the compiler
seems able to optimize things so that music plays smoothly. But with
ledDraw() in led.c, we get warbly audio. So for now, we give the
music app its own privatec copy of this function again since we
want music playback to be smooth.