Add support for temporarily avoiding CPU sleep mode
We use __WFI() in the idle thread to put the CPU to sleep during idle
periods, which saves power. However there's some latency involved in
waking up again. With the music player and bitcoin apps, I sometimes
noticed some warbling which I thought was due to not having enough
spare CPU cycles to handle all the I/O. It looks like the real issue
is that when we put the CPU to sleep, it takes a small amount of time
for it to wake up again, and we just don't have that time to spare here.
The wake-up latency seems to cause us to miss our deadline for getting
new audio samples into the I2S controller, which is where the warble
comes from.

I discovered that this was only noticeable when running the badge
without the JTAG debugger connected. It turns out that when the debugger
is connected, it actually prevents the CPU from going into low power
mode (which makes sense, because the debugger wouldn't be able to
do anything if the CPU core went to sleep). As soon as I plugged the
debugger in, the warble would stop.

To deal with this, the idle thread now uses a custom hook instead of
the built-in ChibiOS __WFI() support. We use a variable to control
whether or not we want to allow a __WFI(). Using badge_sleep_disable()
will clear the variable, and badge_sleep_enable() will turn it back on
again. Right now we disable the sleep function only in the video player
and in the music player app.