Add network game packet protocol structures
These structures have a common header and some context-specific
extra fields. Note that since we want to use the entity type as
a field in the packet header, the entity_type is now a uint16_t
rather than an enum. (The internal representation of enums in C
is Implementation Defined (tm), which is a bad thing when defining
an over-the-wire protocol.)