Copy the config into a 'live' player struct on battle start.
Consoliate player and enemy code in a lot of places so that we're not
having to make special snowflake code for the player. Eeveryone's an
instance of ENEMY now. I hate you. #punk

Advertise the player's ship type along with their rank so we have it
at combat start (and potentially, later, could do custom world_map
icons w/o protocol chnages)