Everything up to the COMBAT screen is now working.
We're not drawing ships yet on that screen (will do tomorrow) but our
clocks work, the VS ship selection is functional, and moving to the
correct map is working. The attacker is in charge of what map we land
on to prevent players being split between maps and proper maps are
being drawn.

Initial combat screen is drawn correctly as well. Nearly there.