Fix a uGFX event processing issue
We only preserve one uGFX event when the uGFX orchard callback is called.
But there are some instances where two events may be generated in rapid
succession by the gtimer thread, and it has higher priority that the
app thread. This means that sometimes an event can be delivered, then
overwritten with another event before the app gets a chance to process
it.

To avoid this race condition, we suspend the thread that calls the
orchard uGFX callback routine until the app has a chance to process
the event.