Remember to reset internal state on exit
"player," "current_enemy," "bullet[x]" and "last_near" need to be
reset to NULL when exiting. player, current_enemy and bullet[x] are
free()d, but that doesn't reset the pointers, and this is not UNIX,
which means our context state is not entirely reset when we re-enter
a task. We have to do it ourselves.

This bug resulted in being unable to properly play a second game,
as we still had a stale current_enemy pointer from the previous one.