Add buffering to the video player
Testing with the production video card showed that video playback was
unacceptably slow. Rewriting the video card with my FreeBSD machine
yielded much better behavior with the same SD card hardware. I suspect
that the data on the master image was written in a less linear fashion
(meaning that files were not always written in nice successive streams
of sectors on disk) which causes lots of seeking when reading back
video files.

We can't re-write all the SD cards though (it would take too long). So
instead, the video player has been refactored to use more of our available
RAM to buffer data from the SD card. We now read 4 video/audio chunks
into RAM instead just reading and playing one chunk at a time. This
yields smooth playback performance. (This was not an option with the DC25
badge due to only having 16KB of RAM to work with, but here we luckily
have 256KB. With the cache changes the video player now uses about 30KB,
which is a bit much, but an acceptable trade-off to get smooth playback.)