Fix memory bugs
In app-battle, remember to release the sprite holders for the
player hit images.

Also, in isp_get_spholder_from_file(), there is a double-free when
the buffer allocated in this function is released. It may already
be released by isp_buf_to_spholder(), so there's no point doing
it again here.