Make thed LED off setting always force LEDs off
In order to work around compatibility problems with certain Shitty
Add-Ons that also use I2C bus address 0x50 (which conflicts with our
LED control chip), we would like to fix it so that we can disable
all I2C bus accesses. Turning the LEDs off helps, but that doesn't
disable LEDs in all circumstances. The music app, geiger counter app
and game app all try to frob the LEDs directly.

To work around this, the "Off" setting now disables all I2C accesses
by the LED code. If you boot with the LEDs in the Off position, this
means the bootstrap messages will report that no LED controller is
found, but changing the setting will result in a re-scan of the bus,
so they will be reactivated correctly.