Fix a possible race condition in the BLE geiger counter app
Wait for the rxAction field to be reset to 0 by the counter thread
before updating it again, otherwise a request from the app thread
might be missed.

On occasion, when programming badges, I found that sometimes the geiger
counter app would not exit correctly (it was still sending packets but
it would stop responding to events). I suspect the problem is that
during exit, setting rxAction to RX_ACTION_EXIT would sometimes get
lost (the thread would overwrite rxAction with a 0 and we'd lose the
event). Hopefully this won't happen anymore.