Check in final provisioning script
Now that DEF CON is over, we can publish the script containing the
final set of unlock code and the unlock password. This is the script
used to initialize all boards actually sold/distributed via Kickstarter,
Stripe and on site at the con.

The secret unlock password is "nucular". (<Homer>"Nucular, Lisa.
Nucular."</Homer>)